:monkey_face: Se agregan los casos del switch y se genera nueva biblioteca inputs.
