+++ tp1/inputs.h
@@ -1,0 +1,23 @@
+#ifndef INPUTS_H_INCLUDED
+#define INPUTS_H_INCLUDED
+
+/** \brief
+ *
+ * \param
+ * \param
+ * \return
+ *
+ */
+
+void clearBuffer(void);
+
+/** \brief
+ *
+ * \param minNumber int
+ * \param maxNumber int
+ * \return int
+ *
+ */
+int getIntOfARange(int minNumber, int maxNumber);
+
+#endif // INPUTS_H_INCLUDED
