+++ tp1/main.c
@@ -32,7 +32,20 @@ int main()
         switch(option)
         {
             case 1:
-                printf("Prueba con caso 1\n");
+                break;
+            case 2:
+                break;
+            case 3:
+                break;
+            case 4:
+                break;
+            case 5:
+                break;
+            case 6:
+                break;
+            case 7:
+                break;
+            case 8:
                 break;
         }
