+++ tp1/menu.h
@@ -3,7 +3,21 @@
 #define OPTION_MIN 1
 #define OPTION_MAX 9
 
+/** \brief
+ *
+ * \param void
+ * \return void
+ *
+ */
 void pauseScreen(void);
+
+/** \brief
+ *
+ * \param x double
+ * \param y double
+ * \return int
+ *
+ */
 int mainMenu(double x, double y);
 
 #endif // MENU_H_INCLUDED
