+++ tp1/tp1.cbp
@@ -19,6 +19,10 @@
 		<Compiler>
 			<Add option="-Wall" />
 		</Compiler>
+		<Unit filename="inputs.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="inputs.h" />
 		<Unit filename="main.c">
 			<Option compilerVar="CC" />
 		</Unit>
